Refactor and use go/db to interact with postgres
