+++ internal/db/schema.go
@@ -1,0 +1,7 @@
+package db
+
+import "github.com/stellar/go/support/db"
+
+type Q struct {
+	db.SessionInterface
+}
